폴더구조, 파일위치를 바꿈
