+++ src/main/java/com/rocket/rocket/controller/HelloController.java
@@ -1,0 +1,14 @@
+package com.rocket.rocket.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class HelloController {
+	
+	@GetMapping("/test")
+	public String test() {
+		return "/test";
+	}
+
+}
