+++ src/main/java/com/rocket/rocket/vo/DoHwVo.java
@@ -1,4 +1,4 @@
-package vo;
+package com.rocket.rocket.vo;
 
 import lombok.Getter;
 import lombok.Setter;
