+++ src/main/java/com/rocket/rocket/vo/NoticeVo.java
@@ -1,23 +1,23 @@
-package vo;
+package com.rocket.rocket.vo;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class ReplyVo {
+public class NoticeVo {
 
-	private String reply_num;
+	private String notice_num;
+	
+	private String title;
 	
 	private String content;
 	
-	private String nickname;
+	private String user_num;
 	
 	private String regDate;
 	
 	private String modiDate;
 	
 	private String delDate;
-	
-	private String comm_num;
 }
