+++ src/main/java/com/rocket/rocket/vo/ReplyVo.java
@@ -1,23 +1,23 @@
-package vo;
+package com.rocket.rocket.vo;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class NoticeVo {
+public class ReplyVo {
 
-	private String notice_num;
-	
-	private String title;
+	private String reply_num;
 	
 	private String content;
 	
-	private String user_num;
+	private String nickname;
 	
 	private String regDate;
 	
 	private String modiDate;
 	
 	private String delDate;
+	
+	private String comm_num;
 }
