--- src/main/java/dao/dasdadakjdhwbdaksjnxakjdnhwqjekn129qw8dhsaijknxm.txt
@@ -1,1 +1,0 @@
-알고보니 이건 긴게 아니였지롱
