+++ src/main/resources/templates/test.mustache
@@ -5,7 +5,6 @@
 <title>Insert title here</title>
 </head>
 <body>
-	<h1>index</h1>
-	<a href="/test">aaa</a>
+	<h1>hello</h1>
 </body>
 </html>
